Move app creation onboarding into its own route
