+++ packages/builder/src/pages/builder/portal/apps/onboarding/_components/DataPanel.svelte
